separate some components, use useSeminarInfo instead of passing seminarId down
