+++ src/components/Problems/Discussion.module.scss
@@ -1,0 +1,52 @@
+.discussionBox{
+  margin: 0px;
+  padding: 0px;
+  height: 100%;
+  width: 100%;
+  display: grid;
+  grid-template-rows: 1fr 8rem;
+  overflow: hidden;
+
+  > div.comments{
+    overflow-y: scroll;
+    overscroll-behavior-y: contain;
+  }
+
+  > div.textArea{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: flex-end;
+
+    textarea{
+      width: calc(100% - 20px);
+      height: 3.4rem;
+      margin: 5px 10px;
+      border: 3px solid black;
+    }
+
+    span{
+      margin: 3px 13px;
+    }
+
+  }
+}
+
+.comment{
+  padding: 10px 5px;
+  margin: 2px 0px;
+
+  > div.title{
+      font-weight: bold;
+      font-style: italic;
+      text-transform: uppercase;
+      font-size: 1.2rem;
+      margin-top: 0px;
+  }
+
+  .commentActions{
+    margin-top: 5px;
+    display: flex;
+    justify-content: flex-end;
+  }
+}
