+++ src/components/Problems/Dropdown.module.scss
@@ -1,0 +1,59 @@
+.dropdown {
+  border-bottom: 5px solid black;
+  width: fit-content;
+  display: flex;
+  align-items: center;
+  text-transform: uppercase;
+  font-weight: bold;
+  font-style: italic;
+  position: relative;
+  margin: 0px 20px;
+
+  .arrow {
+    margin-left: 1rem;
+    display: inline-block;
+    width: 0;
+    height: 0;
+    border-left: 0.5rem solid transparent;
+    border-right: 0.5rem solid transparent;
+
+    border-top: 0.8rem solid black;
+  }
+
+  .options {
+    background-color: white;
+    display: none;
+    position: absolute;
+    width: fit-content;
+    top: 100%;
+    right: 0px;
+    border: 5px solid black;
+    border-width: 5px 5px;
+
+    a {
+      text-decoration: none;
+    }
+  }
+
+  .displayOptions {
+    display: grid;
+    grid-template-columns: max-content;
+  }
+
+  .option {
+    background-color: white;
+    color: black;
+    display: inline-block;
+    white-space: nowrap;
+    width: 100%;
+    padding: 5px;
+  }
+  .option:hover {
+    background-color: #ccc;
+  }
+
+  .option.selectedOption {
+    background-color: black;
+    color: white;
+  }
+}
