+++ src/components/Problems/Problems.module.scss
@@ -64,27 +64,8 @@
     font-weight: bold;
     margin-bottom: 1rem;
   }
-
-  .sideContentContainer{
-    border: 8px solid black;
-    display: grid;
-    grid-template-rows: auto 1fr;
-    height: 100%;
-  }
-
-  .sideContentContainer > div.title{
-    background-color: black;
-    color: white;
-    font-weight: bold;
-    font-style: italic;
-    text-transform: uppercase;
-    font-size: 1.2rem;
-    padding-bottom: 8px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
 }
+
 
 .dropzone{
   display: flex;
@@ -95,69 +76,6 @@
   text-transform: uppercase;
   color: #BBB;
   background-color: #EEE;
-}
-
-.discussionBox{
-    margin: 0px;
-    padding: 0px;
-    height: 100%;
-    width: 100%;
-    display: grid;
-    grid-template-rows: 1fr 8rem;
-    overflow: hidden;
-
-    > div.comments{
-      overflow-y: scroll;
-      overscroll-behavior-y: contain;
-    }
-
-    
-    // > div.textArea:before{
-    //   content: '';
-    //   position: relative;
-    //   top: -10px;
-    //   align-self: center;
-    //   border: 2px solid #DDD;
-    //   width:80%;
-    // }
-
-    > div.textArea{
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: flex-end;
-
-      textarea{
-        width: calc(100% - 20px);
-        height: 3.4rem;
-        margin: 5px 10px;
-        border: 3px solid black;
-      }
-
-      span{
-        margin: 3px 13px;
-      }
-
-    }
-}
-
-.comment{
-    padding: 10px 5px;
-    margin: 2px 0px;
-
-    > div.title{
-        font-weight: bold;
-        font-style: italic;
-        text-transform: uppercase;
-        font-size: 1.2rem;
-        margin-top: 0px;
-    }
-
-    .commentActions{
-      margin-top: 5px;
-      display: flex;
-      justify-content: flex-end;
-    }
 }
 
 .notPublished{
@@ -176,67 +94,7 @@
     display: flex;
     align-items: center;
     justify-content: flex-end;
-  }
-
-  .dropdown {
-    border-bottom: 5px solid black;
-    width: fit-content;
-    display: flex;
-    align-items: center;
-    text-transform: uppercase;
-    font-weight: bold;
-    font-style: italic;
-    position: relative;
-    margin: 0px 20px;
-
-    .arrow {
-      margin-left: 1rem;
-      display: inline-block;
-      width: 0;
-      height: 0;
-      border-left: 0.5rem solid transparent;
-      border-right: 0.5rem solid transparent;
-
-      border-top: 0.8rem solid black;
-    }
-
-    .options {
-      background-color: white;
-      display: none;
-      position: absolute;
-      width: fit-content;
-      top: 100%;
-      right: 0px;
-      border: 5px solid black;
-      border-width: 5px 5px;
-
-      a {
-        text-decoration: none;
-      }
-    }
-
-    .displayOptions {
-      display: grid;
-      grid-template-columns: max-content;
-    }
-
-    .option {
-      background-color: white;
-      color: black;
-      display: inline-block;
-      white-space: nowrap;
-      width: 100%;
-      padding: 5px;
-    }
-    .option:hover {
-      background-color: #ccc;
-    }
-
-    .option.selectedOption {
-      background-color: black;
-      color: white;
-    }
-  }
+}
 
 .problemSubmission{
   border: 1px solid red;
