+++ src/components/Problems/Problems.tsx
@@ -1,15 +1,17 @@
 import axios, {AxiosError} from 'axios'
 import clsx from 'clsx'
-import Link from 'next/link'
 import {useRouter} from 'next/router'
-import {Dispatch, FC, Fragment, MouseEvent, SetStateAction, useEffect, useState} from 'react'
+import {Dispatch, FC, Fragment, SetStateAction, useEffect, useState} from 'react'
 import {useDropzone} from 'react-dropzone'
 
-import {useUser} from '@/utils/UserContext'
+import {UserContainer} from '@/utils/UserContainer'
 import {useSeminarInfo} from '@/utils/useSeminarInfo'
 
 import {Latex} from '../Latex/Latex'
+import {Discussion} from './Discussion'
+import {Dropdown, DropdownOption} from './Dropdown'
 import styles from './Problems.module.scss'
+import {SideContainer} from './SideContainer'
 
 interface SeriesList {
   id: number
@@ -41,7 +43,7 @@ interface Problem {
 }
 
 interface Series {
-  can_pariticipate: boolean // ToDo: spelling - api mistake
+  can_participate: boolean
   is_registered: boolean // ToDo: is_registered should be negated !is_registered - api mistake
   id: number
   problems: Problem[]
@@ -68,25 +70,6 @@ interface Series {
 //   late_tags: string[]
 // }
 
-interface Comments {
-  id: number
-  edit_allowed: boolean
-  text: string
-  posted_at: string
-  published: boolean
-  problem: number
-  posted_by: number
-}
-
-interface Comment {
-  id: number
-  can_edit: boolean
-  text: string
-  published: boolean
-  posted_by: number
-  name: string
-}
-
 interface Profile {
   first_name: string
   last_name: string
@@ -112,11 +95,6 @@ interface Profile {
 //   late_tags: string[]
 // }
 // interfaces for dropdown menus
-interface DropdownOption {
-  id: number
-  text: string
-  link: string
-}
 
 type ProblemsProps = {
   setPageTitle: (title: string) => void
@@ -125,7 +103,7 @@ type ProblemsProps = {
 export const Problems: FC<ProblemsProps> = ({setPageTitle}) => {
   const router = useRouter()
 
-  const user = useUser()
+  const {user} = UserContainer.useContainer()
   // ToDo: initial state false + set value after API update
   const [canRegister, setCanRegister] = useState(true)
   // ToDo: initial state false + set value after API update
@@ -298,15 +276,15 @@ export const Problems: FC<ProblemsProps> = ({setPageTitle}) => {
         setProblems(data.problems)
         setSemesterId(data.semester)
 
-        if (data.can_pariticipate === null) {
+        if (data.can_participate === null) {
           setCanRegister(false)
         } else {
-          setCanRegister(data.can_pariticipate)
+          setCanRegister(data.can_participate)
         }
         if (data.is_registered === null) {
           setRegistered(false)
         } else {
-          setRegistered(!data.is_registered) // ToDo: negate the value - api mistake
+          setRegistered(data.is_registered)
         }
       } catch (e: unknown) {
         const ex = e as AxiosError
@@ -326,7 +304,7 @@ export const Problems: FC<ProblemsProps> = ({setPageTitle}) => {
   const handleRegistrationToSemester = async (id: number) => {
     // ToDo: check user details and use the following request to register the user to semester.
     try {
-      const {data} = await axios.post(`/api/competition/event/${id}/register`)
+      await axios.post(`/api/competition/event/${id}/register`)
     } catch (e: unknown) {
       const ex = e as AxiosError
       const error = ex.response?.status === 404 ? 'Resource not found' : 'An unexpected error has occurred'
@@ -338,7 +316,7 @@ export const Problems: FC<ProblemsProps> = ({setPageTitle}) => {
   return (
     <>
       <div className={styles.container}>
-        <Menu seminarId={seminarId} semesterList={semesterList} selectedId={problemsId} />
+        <Menu semesterList={semesterList} selectedId={problemsId} />
         {problems.map((problem) => (
           <Fragment key={problem.id}>
             <div className={styles.problem}>
@@ -515,15 +493,6 @@ const UploadProblemButton: FC<{
   }
 }
 
-const SideContainer: FC<{title: string}> = ({title, children}) => {
-  return (
-    <div className={styles.sideContentContainer}>
-      <div className={styles.title}>{title}</div>
-      {children}
-    </div>
-  )
-}
-
 const UploadProblemForm: FC<{problemId: number; problemNumber: number}> = ({problemId, problemNumber}) => {
   const {acceptedFiles, getRootProps, getInputProps} = useDropzone()
 
@@ -532,7 +501,7 @@ const UploadProblemForm: FC<{problemId: number; problemNumber: number}> = ({prob
     formData.append('file', acceptedFiles[0])
 
     try {
-      const response = await axios.post(`/api/competition/problem/${problemId}/upload_solution/`, formData)
+      const response = await axios.post(`/api/competition/problem/${problemId}/upload-solution/`, formData)
       if (response.status === 201) {
         console.log('file uploaded') // ToDo: remove log() and let user know the response! message system? or something else?
       }
@@ -567,212 +536,20 @@ const UploadProblemForm: FC<{problemId: number; problemNumber: number}> = ({prob
   // return UploadProblemForm - {problemNumber}
 }
 
-// ToDo: move to the other interfaces
-interface DiscussionProps {
-  problemId: number
-  problemNumber: number
-}
-
-const Discussion: FC<DiscussionProps> = ({problemId, problemNumber}) => {
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState('') // eslint-disable-line @typescript-eslint/no-unused-vars
-  const [comments, setComments] = useState<Comment[]>([])
-  const [commentText, setCommentText] = useState('')
-  // ToDo: find out whether the user can publish comments form the api
-  const [canPublish, setCanPublish] = useState(true)
-  const [reloadComments, setReloadComments] = useState(true)
-  const user = useUser()
-
-  // trigger comments reload whenever problemId changes
-  useEffect(() => {
-    setReloadComments(true)
-  }, [problemId, user])
-
-  // Fetch comments for the problem with id === problemId
-  useEffect(() => {
-    setLoading(true)
-    if (reloadComments === true) {
-      setCommentText('')
-
-      const fetchComments = async () => {
-        try {
-          const {data} = await axios.get<Comments[]>(`/api/competition/problem/${problemId}/comments`, {
-            headers: {
-              'Content-type': 'application/json',
-            },
-          })
-
-          const comments = data.map((comment) => {
-            return {
-              id: comment.id,
-              can_edit: comment.edit_allowed,
-              text: comment.text,
-              published: comment.published,
-              posted_by: comment.posted_by, // ToDo: change after api change
-              name: comment.posted_by.toString(), // ToDo: change after api change
-            }
-          })
-
-          setComments(comments)
-          setLoading(false)
-        } catch (e: unknown) {
-          const ex = e as AxiosError
-          const error = ex.response?.status === 404 ? 'Resource not found' : 'An unexpected error has occurred'
-          setError(error)
-          setComments([])
-        } finally {
-          setLoading(false)
-        }
-      }
-
-      fetchComments()
-      setReloadComments(false)
-    }
-  }, [problemId, reloadComments])
-
-  const handleCommentSubmit = async () => {
-    // ToDo: handle comment submit
-    try {
-      const {data} = await axios.post(`/api/competition/problem/${problemId}/add-comment`, {
-        text: commentText,
-      })
-      setReloadComments(true)
-      // ToDo: use data to show message?
-    } catch (e: unknown) {
-      const ex = e as AxiosError
-      const error = ex.response?.status === 404 ? 'Resource not found' : 'An unexpected error has occurred'
-      // setError(error)
-    } finally {
-      // setLoading(false)
-    }
-  }
-
-  const handleCommentChange = (e: React.FormEvent<HTMLTextAreaElement>) => {
-    setCommentText(e.currentTarget.value)
-  }
-
-  const hideComment = async (id: number) => {
-    try {
-      await axios.post('/api/competition/comment/' + id + '/hide/')
-      setReloadComments(true)
-    } catch (e: unknown) {
-      const ex = e as AxiosError
-      const error = ex.response?.status === 404 ? 'Resource not found' : 'An unexpected error has occurred'
-      // ToDo: handle error
-    }
-  }
-
-  const publishComment = async (id: number) => {
-    try {
-      await axios.post('/api/competition/comment/' + id + '/publish/')
-      setReloadComments(true)
-    } catch (e: unknown) {
-      const ex = e as AxiosError
-      const error = ex.response?.status === 404 ? 'Resource not found' : 'An unexpected error has occurred'
-      // ToDo: handle error
-    }
-  }
-
-  const deleteComment = async (id: number) => {
-    // ToDo: add one extra check before deleting the post
-    // ToDo: use delete method when it is implemented on the backend
-    console.log('delete comment id = ', id)
-    // try {
-    //   await axios.delete('/api/competition/comment/' + id + '/')
-    //   setReloadComments(true)
-    // } catch (e: unknown) {
-    //   const ex = e as AxiosError
-    //   const error = ex.response?.status === 404 ? 'Resource not found' : 'An unexpected error has occurred'
-    //   // ToDo: handle error
-    // }
-  }
-
-  return (
-    <SideContainer title={'Diskusia - úloha ' + problemNumber}>
-      <div className={styles.discussionBox}>
-        <div className={styles.comments}>
-          {/* ToDo: replace by loading component that does not exist yet */}
-          {/* {loading && <div>Loading...</div>} */}
-          {!loading &&
-            comments.map((comment) => {
-              return (
-                <div className={clsx(styles.comment, !comment.published && styles.notPublished)} key={comment.id}>
-                  <div className={styles.title}>
-                    {!comment.published && (
-                      <>
-                        (not published)
-                        <br />
-                      </>
-                    )}
-                    {comment.name}
-                  </div>
-                  <div className={styles.body}>{comment.text}</div>
-                  <div className={styles.commentActions}>
-                    {!comment.published && (
-                      <>
-                        {canPublish && (
-                          <span
-                            onClick={() => {
-                              publishComment(comment.id)
-                            }}
-                          >
-                            Publish
-                          </span>
-                        )}
-                        <span
-                          onClick={() => {
-                            deleteComment(comment.id)
-                          }}
-                        >
-                          Delete
-                        </span>
-                      </>
-                    )}
-                    {comment.published && (
-                      <>
-                        {canPublish && (
-                          <span
-                            onClick={() => {
-                              hideComment(comment.id)
-                            }}
-                          >
-                            Hide
-                          </span>
-                        )}
-                      </>
-                    )}
-                  </div>
-                </div>
-              )
-            })}
-        </div>
-        <div className={styles.textArea}>
-          <textarea value={commentText} onChange={handleCommentChange} />
-          <span onClick={() => handleCommentSubmit()} className={styles.actionButton}>
-            Odoslať
-          </span>
-        </div>
-      </div>
-    </SideContainer>
-  )
-}
-
-const Menu: FC<{seminarId: number; semesterList: SemesterList[]; selectedId: {semester: boolean; id: number}}> = ({
-  seminarId,
+const Menu: FC<{semesterList: SemesterList[]; selectedId: {semester: boolean; id: number}}> = ({
   semesterList,
   selectedId,
 }) => {
+  const {seminar} = useSeminarInfo()
+
   let selectedSemesterId = -1
   let selectedSeriesId = -1
-
-  const [loading, setLoading] = useState(true) // eslint-disable-line @typescript-eslint/no-unused-vars
-  const [error, setError] = useState('') // eslint-disable-line @typescript-eslint/no-unused-vars
 
   const dropdownSemesterList = semesterList.map((semester) => {
     return {
       id: semester.id,
       text: `${semester.year}. Ročník - ${semester.season_code === 0 ? 'zimný' : 'letný'} semester`,
-      link: `/${getSeminarName(seminarId)}/zadania/${semester.year}/${semester.season_code === 0 ? 'zima' : 'leto'}/`,
+      link: `/${seminar}/zadania/${semester.year}/${semester.season_code === 0 ? 'zima' : 'leto'}/`,
     }
   })
 
@@ -785,9 +562,9 @@ const Menu: FC<{seminarId: number; semesterList: SemesterList[]; selectedId: {se
         return {
           id: series.id,
           text: `${series.order}. séria`,
-          link: `/${getSeminarName(seminarId)}/zadania/${semesterList[i].year}/${
-            semesterList[i].season_code === 0 ? 'zima' : 'leto'
-          }/${series.order}/`,
+          link: `/${seminar}/zadania/${semesterList[i].year}/${semesterList[i].season_code === 0 ? 'zima' : 'leto'}/${
+            series.order
+          }/`,
         }
       })
     } else if (selectedId.semester === false) {
@@ -799,7 +576,7 @@ const Menu: FC<{seminarId: number; semesterList: SemesterList[]; selectedId: {se
             return {
               id: series.id,
               text: `${series.order}. séria`,
-              link: `/${getSeminarName(seminarId)}/zadania/${semesterList[i].year}/${
+              link: `/${seminar}/zadania/${semesterList[i].year}/${
                 semesterList[i].season_code === 0 ? 'zima' : 'leto'
               }/${series.order}/`,
             }
@@ -815,53 +592,6 @@ const Menu: FC<{seminarId: number; semesterList: SemesterList[]; selectedId: {se
       <Dropdown title={'Semester'} selectedId={selectedSemesterId} options={dropdownSemesterList} />
     </div>
   )
-}
-
-const Dropdown: FC<{title: string; selectedId: number; options: DropdownOption[]}> = ({title, selectedId, options}) => {
-  const [display, setDisplay] = useState(false)
-
-  const handleClick = (event: MouseEvent) => {
-    event.preventDefault()
-    setDisplay((prevDisplay) => {
-      return !prevDisplay
-    })
-  }
-
-  const handleMouseLeave = () => {
-    setDisplay(false)
-  }
-
-  return (
-    <div className={styles.dropdown} onClick={handleClick} onMouseLeave={handleMouseLeave}>
-      {title} <div className={styles.arrow}></div>
-      <div className={clsx(styles.options, display && styles.displayOptions)}>
-        {options.map((option) => {
-          return (
-            <Link href={option.link} key={option.id}>
-              <a>
-                <div className={clsx(styles.option, selectedId === option.id && styles.selectedOption)}>
-                  {option.text}
-                </div>
-              </a>
-            </Link>
-          )
-        })}
-      </div>
-    </div>
-  )
-}
-
-const getSeminarName = (id: number) => {
-  switch (id) {
-    case 0:
-      return 'strom'
-    case 1:
-      return 'matik'
-    case 2:
-      return 'malynar'
-    default:
-      return ''
-  }
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
