+++ src/components/Problems/SideContainer.module.scss
@@ -1,0 +1,19 @@
+.sideContentContainer {
+  border: 8px solid black;
+  display: grid;
+  grid-template-rows: auto 1fr;
+  height: 100%;
+}
+
+.sideContentContainer > div.title {
+  background-color: black;
+  color: white;
+  font-weight: bold;
+  font-style: italic;
+  text-transform: uppercase;
+  font-size: 1.2rem;
+  padding-bottom: 8px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
